bot/requirements.txt
# Bot de Trading - Dependencias Python

# MetaTrader5
MetaTrader5>=5.0.45

# Telegram Client (leer señales de canales)
telethon>=1.34.0

# Telegram Bot (notificaciones y comandos)
python-telegram-bot>=21.0

# HTTP requests
requests>=2.31.0
urllib3>=2.0.0

# Opcional: psutil para métricas de sistema
psutil>=5.9.0
